pref: divide `get.is`.
